requirements.txt
aniso8601==9.0.1
apispec==5.2.2
apispec-webframeworks==0.5.2
attrs==21.4.0
cachelib==0.9.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
flask-apispec==0.11.0
Flask-Caching==2.0.0
Flask-Cors==3.0.10
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.8.0
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-mongoengine==0.31.0
mongoengine==0.24.2
packaging==21.3
pymongo==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
PyYAML==6.0
six==1.16.0
ua-parser-py==1.0.2
webargs==8.2.0
Werkzeug==2.2.1
WTForms==3.0.1
locust==2.10.2
autotech-sdk==1.0.2
python-dotenv==0.21.0
